refactor: removed dead code at star.py
