+++ python/bloqade/lanes/dialects/place.py
@@ -6,6 +6,7 @@ from kirin.dialects import ilist
 from kirin.lattice.empty import EmptyLattice
 
 from bloqade import types as bloqade_types
+from bloqade.gemini.star import validate_steane_star_support
 from bloqade.lanes.analysis.layout import LayoutAnalysis
 from bloqade.lanes.analysis.placement import (
     AtomState,
@@ -14,7 +15,6 @@ from bloqade.lanes.analysis.placement import (
     PlacementAnalysis,
 )
 from bloqade.lanes.bytecode.encoding import LocationAddress
-from bloqade.lanes.star import validate_steane_star_support
 from bloqade.lanes.types import StateType
 
 dialect = ir.Dialect(name="lanes.place")
